histogram

+++ src/main/FreeEvent.java
@@ -3,10 +3,10 @@ package main;
 import memory.MemoryBlock;
 
 public class FreeEvent {
-	int time;
+	long time;
 	MemoryBlock block;
 	
-	public FreeEvent(int d, MemoryBlock block) {
+	public FreeEvent(long d, MemoryBlock block) {
 		super();
 		this.time = d;
 		this.block = block;
